01_ai_agents_first/17a_session_memory/pyproject.toml
[project]
name = "17a-session-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.3",
]
